Get rid of warning "warning: comparison is always true due to limited range of data type" caused by the fact that strtol returns a long, but is passed an integer (and integer can't hold LONG_MAX/LONG_MIN)

+++ src/util.c
@@ -156,7 +156,7 @@ util_char_to_arg(char *value, DwbType type) {
       }
     }
     else if (type == INTEGER) {
-      int n = strtol(value, NULL, 10);
+      long n = strtol(value, NULL, 10);
       if (n != LONG_MAX &&  n != LONG_MIN && !errno ) {
         ret->i = n;
       }
